+++ Changelog.md
@@ -24,6 +24,7 @@
 - Fix `extra_hosts` configuration in `compose.yaml` (#7375)
 - Add unit tests for `marks_graders_controller` (#7382)
 - Convert front-end tests from enzyme to react testing library; add `@testing-library/user-event` (#7379)
+- Refactor the `Result` component and its children to use React context API (#7380)
 
 ## [v2.6.1]